Remove AWS4AuthLayer and HTTP.request extension
It's currently exceedingly difficult for packages to extend
`HTTP.request` by providing a custom layer, which is likely why the
logic for an AWS authentication layer is in HTTP rather than AWSCore.
The simplest course of action would be to remove the functionality from
this package and instead introduce a dependency on AWSAuth to HTTP, with
the `request` extension using signing defined here.